+++ RELEASES.md
@@ -1,6 +1,7 @@
 
 * [ONEAPP-15666](https://jira.tau2904.com/browse/ONEAPP-15666) [BE] Service /products/credit-card/statement/get-unbilled-statement working to not correct
 * [ONEAPP-16586](https://jira.tau2904.com/browse/ONEAPP-16586) [BE] Create initial vector param field
+* [ONEAPP-18412](https://jira.tau2904.com/browse/ONEAPP-18412) [Apply eStatement] After finish success apply eStatement C2G - menu still display
 * [ONEAPP-21818](https://jira.tau2904.com/browse/ONEAPP-21818) [Set PIN] Activity Log display incorrect
 * [ONEAPP-21897](https://jira.tau2904.com/browse/ONEAPP-21897) (BE) Apply eStatement - require clearing radish after success applying
 * [ONEAPP-22137](https://jira.tau2904.com/browse/ONEAPP-22137) Create property for cal risk
@@ -12,6 +13,7 @@
 * [ONEAPP-23184](https://jira.tau2904.com/browse/ONEAPP-23184) FiID
 * [ONEAPP-23187](https://jira.tau2904.com/browse/ONEAPP-23187) Edit transaction fee for selling from frontEndFee to BackEndFee
 * [ONEAPP-23546](https://jira.tau2904.com/browse/ONEAPP-23546) [Apply so good] Act log apply so good fail case display incorrect
+* [ONEAPP-24397](https://jira.tau2904.com/browse/ONEAPP-24397) [Apply e Statement] Cant do Apply e statement
 * [ONEAPP-13008](https://jira.tau2904.com/browse/ONEAPP-13008) (BE) Submit uploading
 * [ONEAPP-14375](https://jira.tau2904.com/browse/ONEAPP-14375) Add credit card for buy flow .
 * [ONEAPP-14532](https://jira.tau2904.com/browse/ONEAPP-14532) [Personal information]Incase of life time citizen id ,Expiry date of citizen card display incorrect